Move Confirm pledge tests to full form flow & fix bug exposed
